改造

+++ fast-data-switch/fast-data-switch-admin/src/main/java/cn/zhumingwu/dataswitch/admin/repository/JobInfoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobInfoRepository extends JpaRepository<JobInfo, Integer> {
+public interface JobInfoRepository extends JpaRepository<JobInfo, Long> {
 }
